@ingrid-kis [Task: Final Update Blog Pessoal]

+++ src/main/java/org/generation/blogPessoal/controller/TemaController.java
@@ -1,6 +1,9 @@
 package org.generation.blogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.generation.blogPessoal.model.Tema;
 import org.generation.blogPessoal.repository.TemaRepository;
@@ -16,6 +19,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -50,13 +54,14 @@ public class TemaController {
 		}) 
 		@GetMapping("/{id}")
 		public ResponseEntity<Tema> getById(@PathVariable long id){
-			return repository.findById(id).map(resp -> ResponseEntity.ok(resp))
-					.orElse(ResponseEntity.notFound().build());
+			return repository.findById(id)
+					.map(resp -> ResponseEntity.ok(resp))
+					.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 		}
 		
-		@GetMapping("/nome/{nome}")
-		public ResponseEntity<List<Tema>> getByName(@PathVariable String nome){
-			return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(nome));
+		@GetMapping("/descricao/{descricao}")
+		public ResponseEntity<List<Tema>> getByTitulo(@PathVariable String descricao){
+			return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(descricao));
 		}
 		
 	    @Operation(summary = "Cria novo tema")
@@ -67,7 +72,7 @@ public class TemaController {
 	            @ApiResponse(responseCode = "500", description = "Erro interno no servidor")
 		})
 		@PostMapping
-		public ResponseEntity<Tema> post (@RequestBody Tema tema){
+		public ResponseEntity<Tema> post (@Valid @RequestBody Tema tema){
 			return ResponseEntity.status(HttpStatus.CREATED)
 					.body(repository.save(tema));
 		}
@@ -79,8 +84,10 @@ public class TemaController {
 	            @ApiResponse(responseCode = "500", description = "Erro interno no servidor")
 		})
 		@PutMapping
-		public ResponseEntity<Tema> put (@RequestBody Tema tema){
-			return ResponseEntity.ok(repository.save(tema));
+		public ResponseEntity<Tema> put(@Valid @RequestBody Tema tema){
+			return repository.findById(tema.getId())
+					.map(resp -> ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema)))
+					.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 		}
 		
 	    @Operation(summary = "Deleta tema existente")
@@ -90,6 +97,12 @@ public class TemaController {
 		})
 		@DeleteMapping("/{id}")
 		public void delete(@PathVariable long id) {
+	    	Optional<Tema> tema = repository.findById(id);
+
+			if(tema.isEmpty())
+				throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
+	    	
 			repository.deleteById(id);
 		}
 }
